fix: Move TaskFunc to client side
