+++ scheduler/src/task.rs
@@ -3,18 +3,7 @@ use std::time::Duration;
 use chrono::{offset::Utc, DateTime};
 use serde::{Deserialize, Serialize};
 
-use super::{ResourceAlloc, ResourceReq};
-
-pub trait TaskFunc {
-    type Output;
-    type Error;
-
-    fn init(&mut self, _: Option<&ResourceAlloc>) -> Result<(), Self::Error> {
-        Ok(())
-    }
-    fn end(&mut self, _: Option<&ResourceAlloc>) -> Result<Self::Output, Self::Error>;
-    fn task(&mut self, alloc: Option<&ResourceAlloc>) -> Result<TaskResult, Self::Error>;
-}
+use super::ResourceReq;
 
 /// Helper type that indicates if a task should be executed again
 #[derive(PartialEq, Eq)]
